64bits

+++ json_object.c
@@ -642,7 +642,7 @@ struct json_object* json_object_new_string(const char *s)
   return jso;
 }
 
-struct json_object* json_object_new_string_len(const char *s, int len)
+struct json_object* json_object_new_string_len(const char *s, size_t len)
 {
   struct json_object *jso = json_object_new(json_type_string);
   if(!jso) return NULL;
